+++ cypress/cypress/component/Typeahead/AsyncTypeaheadInput.cy.tsx
@@ -650,9 +650,6 @@ it("placeholder", () => {
         onSubmit={() => {
           // Nothing to do
         }}
-        defaultValues={{
-          [name]: simpleOptions,
-        }}
       >
         <AsyncTypeaheadInput
           multiple
@@ -671,7 +668,10 @@ it("placeholder", () => {
       </Form>
     </div>,
   );
+
   cy.get(`#${name}`).should("have.attr", "placeholder", placeholder);
+  simpleOptions.slice(0, 2).forEach((option) => selectOption(name, option));
+  cy.get(`#${name}`).should("not.have.attr", "placeholder");
 });
 
 it("test on input change", () => {