+++ src/lib/FormattedInput.tsx
@@ -10,6 +10,7 @@ import { useFormContext } from "./context/FormContext";
 interface FormattedInputProps<T extends FieldValues> extends CommonInputProps<T> {
   patternFormat?: PatternFormatProps;
   numericFormat?: NumericFormatProps;
+  placeholder?: string;
 }
 
 const FormattedInput = <T extends FieldValues>(props: FormattedInputProps<T>) => {
@@ -33,6 +34,7 @@ const FormattedInput = <T extends FieldValues>(props: FormattedInputProps<T>) =>
     addonRight,
     className = "",
     hideValidationMessage = false,
+    placeholder,
   } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
   const { control, disabled: formDisabled } = useFormContext();
@@ -89,6 +91,7 @@ const FormattedInput = <T extends FieldValues>(props: FormattedInputProps<T>) =>
                   }}
                   onFocus={focusHandler}
                   style={style}
+                  placeholder={placeholder}
                 ></NumericFormat>
               )}