+++ CHANGELOG.md
@@ -11,6 +11,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - migrated to eslint 9 flat config
 
+## [3.3.0] - 2025-06-03
+
+### Added
+
+- property `placeholder` to `FormattedInput` component.
+
+## [3.2.0] - 2025-04-30
+
+### Added
+
+- Export of typeahead helpers and `useDebounceHook` hook.
+
+### Fixed
+
+- Hide placeholder on multiple `AsyncTypeAheadInput` and `StaticTypeAheadInput` when at least one option is selected.
+- Typeahead helpers return type from `TypeaheadOption[]` to `TypeaheadOptions`, since typeaheads are not intended to be used with mixed `string` and `LabelValueOption` options.
+
 ## [3.1.1] - 2025-04-10
 
 ### Fixed
@@ -492,7 +509,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/3.1.1...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/3.3.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -517,6 +534,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[3.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/3.2.0...3.3.0
+[3.2.0]: https://github.com/neolution-ch/react-hook-form-components/compare/3.1.1...3.2.0
 [3.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/3.1.0...3.1.1
 [3.1.0]: https://github.com/neolution-ch/react-hook-form-components/compare/3.0.1...3.1.0
 [3.0.1]: https://github.com/neolution-ch/react-hook-form-components/compare/3.0.0...3.0.1